refactor(front): merge stringHelper and textHelper
removes duplicate
updates imports
