+++ packages/front/svelte/helpers/stringHelper.js
@@ -7,3 +7,7 @@ export function trim(str, length) {
 export function isHyphen(str) {
     return str === "-";
 }
+
+export function capitalizeFirstLetter(str) {
+    return str ? str[0].toUpperCase() + str.slice(1).toLowerCase() : str;
+}
